refactor: remove orchestrator abstraction from API
