+++ examples/randomwalks/configs/ppo_randomwalks.yml
@@ -8,7 +8,6 @@ train:
   eval_interval: 20
 
   pipeline: "PromptPipeline"
-  orchestrator: "PPOOrchestrator"
   trainer: "AcceleratePPOTrainer"
 
 model:
