+++ examples/summarize_daily_cnn/configs/ppo_config_cnn_daily.yml
@@ -9,7 +9,6 @@ train:
   save_best: False
 
   pipeline: "PromptPipeline"
-  orchestrator: "PPOOrchestrator"
   trainer: "AcceleratePPOTrainer"
 
 model:
